+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 # Unreleased
 
-> NOTE: starting from this release, you can optionally use the new worker by
-> adding the `--worker-class fractal_server.worker.FractalWorker` option to
-> the `gunicorn` startup command.
+> NOTE: you can enable custom gunicorn worker/logger by adding the following options to the `gunicorn` startup command:
+> - `--worker-class fractal_server.gunicorn_fractal.FractalWorker`
+> - `--logger-class fractal_server.gunicorn_fractal.FractalGunicornLogger`
 
+
+* App:
+    * Move `FractalGunicornLogger` and `FractalWorker` in `fractal_server/gunicorn_fractal.py` (\#1535).
 * API:
     * Remove catch of `IntegrityError` in `POST /api/v1/project` (\#1530).
     * Add extensive logs to `DELETE /api/v2/project/{project_id}` (\#1532).